--- lib/composition-provider/select-files.js
@@ -1,66 +1,0 @@
-// This is kind of META provider, invoked from UI.
-// Provide narrow-ui to select narrowing target files.
-const {Point} = require('atom')
-const Provider = require('./provider')
-const settings = require('../settings')
-const {relativizeFilePath} = require('../utils')
-
-const queryByProviderName = {}
-
-module.exports = class SelectFiles {
-  constructor (clientUi) {
-    this.clientUi = clientUi
-
-    this.provider = Provider.create({
-      name: this.constructor.name,
-      config: {
-        boundToSingleFile: true,
-        supportCacheItems: true,
-        supportReopen: false,
-        needRestoreEditorState: false
-      },
-      onDestroyed: () => {
-        if (this.clientUi.isAlive()) {
-          this.clientUi.focus({autoPreview: false})
-        }
-      },
-      onItemConfirmed: this.onItemConfirmed.bind(this),
-      getItems: () => this.getItems()
-    })
-  }
-
-  start (options) {
-    this.provider.start(options)
-  }
-
-  static getLastQuery (providerName) {
-    let query
-    if (settings.get('SelectFiles.rememberQuery')) {
-      const query = queryByProviderName[providerName]
-    }
-    return query || ''
-  }
-
-  getItems () {
-    const filePaths = this.clientUi.getFilePathsForAllItems()
-    this.provider.finishUpdateItems(
-      filePaths.map(filePath => ({
-        text: relativizeFilePath(filePath),
-        filePath,
-        point: new Point(0, 0)
-      }))
-    )
-  }
-
-  // Must return undefined, when non null value was returned
-  // Ui treat it as openable editor instance and try to open it.
-  async onItemConfirmed (item) {
-    if (this.clientUi.isAlive()) {
-      const queryForSelectFiles = this.provider.ui.lastQuery
-      await this.clientUi.resetQueryForSelectFiles(queryForSelectFiles)
-      this.clientUi.moveToItemForFilePath(item.filePath)
-      queryByProviderName[this.clientUi.provider.name] = queryForSelectFiles
-    }
-    this.provider.ui.destroy()
-  }
-}
