+++ lib/main.js
@@ -2,8 +2,7 @@ const {CompositeDisposable, Disposable} = require('atom')
 const settings = require('./settings')
 const Ui = require('./ui')
 const globalSubscriptions = require('./global-subscriptions')
-const ProviderBase = require('./provider/provider-base')
-const Provider = require('./composition-provider/provider')
+const Provider = require('./provider/provider')
 
 const {isNarrowEditor, getVisibleEditors, isTextEditor, suppressEvent} = require('./utils')
 module.exports = {
@@ -109,32 +108,7 @@ module.exports = {
       'narrow:atom-scan': () => this.narrow('atom-scan'),
       'narrow:atom-scan-by-current-word': () => this.narrow('atom-scan', {queryCurrentWord: true}),
 
-      'narrow:toggle-search-start-by-double-click': () => settings.toggle('Search.startByDoubleClick'),
-
-      // ==== composition based
-      'narrow:x-symbols': () => this.startNarrow('symbols'),
-      'narrow:x-symbols-by-current-word': () => this.startNarrow('symbols', {queryCurrentWord: true}),
-
-      'narrow:x-project-symbols': () => this.startNarrow('project-symbols'),
-      'narrow:x-project-symbols-by-current-word': () => this.startNarrow('project-symbols', {queryCurrentWord: true}),
-
-      'narrow:x-git-diff-all': () => this.startNarrow('git-diff-all'),
-
-      'narrow:x-fold': () => this.startNarrow('fold'),
-      'narrow:x-fold-by-current-word': () => this.startNarrow('fold', {queryCurrentWord: true}),
-
-      'narrow:x-scan': () => this.startNarrow('scan'),
-      'narrow:x-scan-by-current-word': () => this.startNarrow('scan', {queryCurrentWord: true}),
-
-      'narrow:x-search': () => this.startNarrow('search'),
-      'narrow:x-search-by-current-word': () => this.startNarrow('search', {queryCurrentWord: true}),
-      'narrow:x-search-by-current-word-without-focus': () => this.startNarrow('search', {queryCurrentWord: true, focus: false}),
-      'narrow:x-search-current-project': () => this.startNarrow('search', {currentProject: true}),
-      'narrow:x-search-current-project-by-current-word': () => this.startNarrow('search', {currentProject: true, queryCurrentWord: true}),
-
-      'narrow:x-atom-scan': () => this.startNarrow('atom-scan'),
-      'narrow:x-atom-scan-by-current-word': () => this.startNarrow('atom-scan', {queryCurrentWord: true})
-      // ==== composition based
+      'narrow:toggle-search-start-by-double-click': () => settings.toggle('Search.startByDoubleClick')
     })
   },
 
@@ -188,15 +162,11 @@ module.exports = {
   },
 
   reopen () {
-    ProviderBase.reopen()
+    Provider.reopen()
   },
 
-  narrow (...args) {
-    ProviderBase.start(...args)
-  },
-
-  startNarrow (name, options) {
-    const Klass = require(`./composition-provider/${name}`)
+  narrow (name, options) {
+    const Klass = require(`./provider/${name}`)
     new Klass().start(options)
   },
 
@@ -239,17 +209,17 @@ module.exports = {
   },
 
   consumeInlineGitDiff (service) {
-    ProviderBase.setService({inlineGitDiff: service})
     Provider.setService({inlineGitDiff: service})
   },
 
   registerProvider (name, klassOrFilePath) {
-    ProviderBase.registerProvider(name, klassOrFilePath)
+    Provider.registerProvider(name, klassOrFilePath)
   },
 
   provideNarrow () {
     return {
-      ProviderBase,
+      ProviderBase: Provider,
+      Provider: Provider,
       registerProvider: this.registerProvider.bind(this),
       narrow: this.narrow.bind(this)
     }
