+++ lib/provider/atom-scan.js
@@ -1,10 +1,10 @@
 'use babel'
 
 const {Range} = require('atom')
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const {scanItemsForFilePath} = require('../utils')
 
-const providerConfig = {
+const Config = {
   supportDirectEdit: false,
   showColumnOnLineHeader: true,
   showProjectHeader: true,
@@ -17,57 +17,69 @@ const providerConfig = {
   refreshOnDidStopChanging: true
 }
 
-module.exports = class AtomScan extends ProviderBase {
-  constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
-    this.itemizeResult = this.itemizeResult.bind(this)
+function createItemsFromScanResult ({filePath, matches}) {
+  return matches.map(({range, lineText, lineTextOffset}) => {
+    range = Range.fromObject(range)
+    return {
+      filePath: filePath,
+      text: lineText,
+      point: range.start,
+      range: range,
+      translateRange () {
+        return this.range.translate([0, -lineTextOffset])
+      }
+    }
+  })
+}
+
+module.exports = class AtomScan {
+  constructor () {
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: Config,
+      getItems: this.getItems.bind(this),
+      onDestroyed: () => {
+        if (this.scanPromise) {
+          this.scanPromise.cancel()
+          this.scanPromise = null
+        }
+      }
+    })
   }
 
-  itemizeResult (result) {
-    if (result && result.matches && result.matches.length) {
-      const {filePath, matches} = result
-      const items = matches.map(({range, lineText, lineTextOffset}) => {
-        range = Range.fromObject(range)
-        return {
-          filePath: filePath,
-          text: lineText,
-          point: range.start,
-          range: range,
-          translateRange () {
-            return this.range.translate([0, -lineTextOffset])
-          }
-        }
-      })
-      this.updateItems(items)
-    }
+  start (options) {
+    this.provider.start(options)
   }
 
   async getItems ({filePath}) {
     if (this.scanPromise) this.scanPromise.cancel()
 
-    const {searchRegex} = this.searchOptions
-
+    const {searchRegex} = this.provider.searchOptions
     if (!searchRegex) {
-      return this.finishUpdateItems([])
+      return this.provider.finishUpdateItems([])
     }
 
     if (filePath) {
       if (atom.project.contains(filePath)) {
-        this.finishUpdateItems(await scanItemsForFilePath(filePath, searchRegex))
+        this.provider.finishUpdateItems(await scanItemsForFilePath(filePath, searchRegex))
       } else {
-        this.finishUpdateItems([])
+        this.provider.finishUpdateItems([])
       }
     } else {
-      this.scanPromise = atom.workspace.scan(searchRegex, this.itemizeResult).then(message => {
-        this.scanPromise = null
+      this.scanPromise = atom.workspace.scan(searchRegex, result => {
+        if (result && result.matches && result.matches.length) {
+          this.provider.updateItems(createItemsFromScanResult(result))
+        }
+      })
+
+      const message = await this.scanPromise
+      this.scanPromise = null
+      if (message !== 'cancelled') {
         // Relying on Atom's workspace.scan's specific implementation
         // `workspace.scan` return cancellable promise.
         // When cancelled, promise is NOT rejected, instead it's resolved with 'cancelled' message
-        if (message !== 'cancelled') {
-          this.finishUpdateItems()
-        }
-      })
+        this.provider.finishUpdateItems()
+      }
     }
   }
 }
