+++ lib/provider/fold.js
@@ -1,4 +1,5 @@
-const ProviderBase = require('./provider-base')
+const settings = require('../settings')
+const Provider = require('./provider')
 const {limitNumber, getFirstCharacterPositionForBufferRow} = require('../utils')
 
 function getCodeFoldStartRows (editor, indentLevel) {
@@ -10,36 +11,46 @@ function getCodeFoldStartRows (editor, indentLevel) {
     .filter(row => editor.indentationForBufferRow(row) < indentLevel)
 }
 
-module.exports = class Fold extends ProviderBase {
-  constructor (...args) {
-    super(...args)
-    Object.assign(this, {
-      boundToSingleFile: true,
-      foldLevel: this.getConfig('foldLevel'),
-      supportCacheItems: true,
-      refreshOnDidStopChanging: true
+module.exports = class Fold {
+  constructor () {
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        boundToSingleFile: true,
+        foldLevel: settings.get('Fold.foldLevel'),
+        supportCacheItems: true,
+        refreshOnDidStopChanging: true
+      },
+      onUiCreated: ui => {
+        atom.commands.add(ui.editor.element, {
+          'narrow-ui:fold-increase-level': () => this.updateFoldLevel(+1),
+          'narrow-ui:fold-decrease-level': () => this.updateFoldLevel(-1)
+        })
+      },
+      getItems: () => this.getItems()
     })
   }
 
-  initialize () {
-    atom.commands.add(this.ui.editorElement, {
-      'narrow-ui:fold-increase-level': () => this.updateFoldLevel(+1),
-      'narrow-ui:fold-decrease-level': () => this.updateFoldLevel(-1)
-    })
+  start (options) {
+    this.provider.start(options)
   }
 
   updateFoldLevel (relativeLevel) {
-    this.foldLevel = limitNumber(this.foldLevel + relativeLevel, {min: 1})
-    this.setConfig('foldLevel', this.foldLevel)
-    return this.ui.refresh({force: true})
+    const oldValue = settings.get('Fold.foldLevel')
+    const newValue = limitNumber(oldValue + relativeLevel, {min: 1})
+    if (newValue !== oldValue) {
+      settings.set('Fold.foldLevel', newValue)
+      this.provider.ui.refresh({force: true})
+    }
   }
 
   getItems () {
-    const foldStartRows = getCodeFoldStartRows(this.editor, this.foldLevel)
-    this.finishUpdateItems(
+    const editor = this.provider.editor
+    const foldStartRows = getCodeFoldStartRows(editor, settings.get('Fold.foldLevel'))
+    this.provider.finishUpdateItems(
       foldStartRows.map(row => ({
-        point: getFirstCharacterPositionForBufferRow(this.editor, row),
-        text: this.editor.lineTextForBufferRow(row)
+        point: getFirstCharacterPositionForBufferRow(editor, row),
+        text: editor.lineTextForBufferRow(row)
       }))
     )
   }
