+++ lib/provider/git-diff-all.js
@@ -1,10 +1,11 @@
 'use babel'
 
 const {Point, Disposable} = require('atom')
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const Path = require('path')
 const fs = require('fs-plus')
 const {limitNumber} = require('../utils')
+const settings = require('../settings')
 
 function getModifiedFilePathsForRepository (repository) {
   const dir = repository.getWorkingDirectory()
@@ -45,43 +46,47 @@ function repositoryForPath (filePath) {
   }
 }
 
-const providerConfig = {
-  showProjectHeader: true,
-  showFileHeader: true,
-  supportCacheItems: true,
-  supportFilePathOnlyItemsUpdate: true,
-  refreshOnDidStopChanging: true
-}
-
-module.exports = class GitDiffAll extends ProviderBase {
+module.exports = class GitDiffAll {
   constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
     this.enabledInlineGitDiff = new Set()
-    this.subscriptions.add(new Disposable(() => this.destroyEnabledInlineGitDiff()))
-    this.toggleInlineGitDiff = this.toggleInlineGitDiff.bind(this)
-  }
 
-  initialize () {
-    atom.commands.add(this.ui.editorElement, {
-      'narrow-ui:git-diff-all-toggle-inline-diff': () => this.toggleInlineGitDiff()
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        showProjectHeader: true,
+        showFileHeader: true,
+        supportCacheItems: true,
+        supportFilePathOnlyItemsUpdate: true,
+        refreshOnDidStopChanging: true
+      },
+      onUiCreated: ui => {
+        atom.commands.add(ui.editor.element, {
+          'narrow-ui:git-diff-all-toggle-inline-diff': () => this.toggleInlineGitDiff()
+        })
+      },
+      onItemOpened: editor => {
+        if (settings.get('GitDiffAll.inlineGitDiffIntegration')) {
+          if (this.lastOpenedEditor !== editor) {
+            this.lastOpenedEditor = editor
+            this.enableInlineGitDiffForEditor(editor)
+          }
+        }
+      },
+      onDestroyed: this.destroyEnabledInlineGitDiff.bind(this),
+      getItems: this.getItems.bind(this)
     })
   }
 
-  onItemOpened (editor) {
-    if (this.getConfig('inlineGitDiffIntegration')) {
-      if (this.lastOpenedEditor !== editor) {
-        this.lastOpenedEditor = editor
-        this.enabledInlineGitDiffForEditor(editor)
-      }
-    }
+  start (options) {
+    this.provider.start(options)
   }
 
-  enabledInlineGitDiffForEditor (editor) {
-    if (this.service.inlineGitDiff) {
-      const inlineGitDiff = this.service.inlineGitDiff.getInlineGitDiff(editor)
-      // If inlineGitDiff was NOT already enabled we dispose it on narrow:close
+  enableInlineGitDiffForEditor (editor) {
+    if (Provider.service.inlineGitDiff) {
+      const inlineGitDiff = Provider.service.inlineGitDiff.getInlineGitDiff(editor)
+      // enable() return `false` if it's already enabled
       if (inlineGitDiff.enable()) {
+        // if we enabled it, remember to disable on `narrow:close`
         this.enabledInlineGitDiff.add(inlineGitDiff)
       }
     }
@@ -97,42 +102,40 @@ module.exports = class GitDiffAll extends ProviderBase {
       require('atom-package-deps').install('narrow')
     }
 
-    const enabled = !this.getConfig('inlineGitDiffIntegration')
-    this.setConfig('inlineGitDiffIntegration', enabled)
+    const enabled = !settings.get('GitDiffAll.inlineGitDiffIntegration')
+    settings.set('GitDiffAll.inlineGitDiffIntegration', enabled)
     if (enabled) {
-      this.enabledInlineGitDiffForEditor(this.lastOpenedEditor || this.editor)
+      this.enableInlineGitDiffForEditor(this.lastOpenedEditor || this.provider.editor)
     } else {
       this.destroyEnabledInlineGitDiff()
     }
     this.updateControlBar(enabled)
   }
 
-  updateControlBar (value = this.getConfig('inlineGitDiffIntegration')) {
-    this.ui.controlBar.updateElements({inlineGitDiff: value})
+  updateControlBar (value = settings.get('GitDiffAll.inlineGitDiffIntegration')) {
+    this.provider.ui.controlBar.updateElements({inlineGitDiff: value})
   }
 
-  getItems ({filePath}) {
+  async getItems ({filePath}) {
     this.updateControlBar()
 
     if (filePath) {
       const repo = repositoryForPath(filePath)
-      if (repo) {
-        getItemsForFilePath(repo, filePath).then(this.finishUpdateItems)
-      } else {
-        this.finishUpdateItems([])
-      }
+      const items = repo ? await getItemsForFilePath(repo, filePath) : []
+      this.provider.finishUpdateItems(items)
     } else {
       const promises = []
-      const updateItems = items => this.updateItems(items.filter(v => v))
+      const updateItems = items => this.provider.updateItems(items.filter(v => v))
 
-      const repositories = atom.project.getRepositories().filter(repo => repo != null)
+      const repositories = atom.project.getRepositories().filter(repo => repo)
       for (const repo of repositories) {
         for (const filePath of getModifiedFilePathsForRepository(repo)) {
           const promise = getItemsForFilePath(repo, filePath).then(updateItems)
           promises.push(promise)
         }
       }
-      Promise.all(promises).then(() => this.finishUpdateItems())
+      await Promise.all(promises)
+      this.provider.finishUpdateItems()
     }
   }
 }
