+++ lib/provider/project-symbols.js
@@ -9,7 +9,7 @@ const {CompositeDisposable, File, Disposable} = require('atom')
 // - Greately depending on core-package `symbols-view`
 // - Since this provider bollowing many function from `symbols-view`.
 
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const {requireFrom, compareByPoint} = require('../utils')
 const globalSubscriptions = require('../global-subscriptions')
 
@@ -36,9 +36,11 @@ function watchTagsFiles () {
     const tagsFilePath = getTagsFile(projectPath)
     if (tagsFilePath) {
       const tagsFile = new File(tagsFilePath)
-      watchSubscriptions.add(tagsFile.onDidChange(clearCachedItems))
-      watchSubscriptions.add(tagsFile.onDidDelete(clearCachedItems))
-      watchSubscriptions.add(tagsFile.onDidRename(clearCachedItems))
+      watchSubscriptions.add(
+        tagsFile.onDidChange(clearCachedItems),
+        tagsFile.onDidDelete(clearCachedItems),
+        tagsFile.onDidRename(clearCachedItems)
+      )
     }
   }
 }
@@ -71,35 +73,41 @@ const itemForTag = function (tag) {
   }
 }
 
-const providerConfig = {
-  showProjectHeader: true,
-  showFileHeader: true,
-  queryWordBoundaryOnByCurrentWordInvocation: true,
-  refreshOnDidSave: true
+function refreshPointForFilePath (event) {
+  const items = getCachedItems()
+  if (items) {
+    const filePath = event.path
+    items.filter(item => item.filePath === filePath).forEach(item => item.refreshPoint())
+  }
 }
 
-module.exports = class ProjectSymbols extends ProviderBase {
+module.exports = class ProjectSymbols {
   constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        showProjectHeader: true,
+        showFileHeader: true,
+        queryWordBoundaryOnByCurrentWordInvocation: true,
+        refreshOnDidSave: true
+      },
+      onUiCreated: ui => {
+        // When user manually refresh, clear cache
+        this.provider.subscriptions.add(ui.onWillRefreshManually(clearCachedItems))
+      },
+      onBindEditor: this.onBindEditor.bind(this),
+      onDestroyed: this.stop.bind(this),
+      getItems: this.getItems.bind(this)
+    })
   }
 
   onBindEditor ({newEditor}) {
     // Refresh item.point in cachedItems for saved filePath.
-    this.subscribeEditor(newEditor.onDidSave(refreshPointForFilePath))
-
-    function refreshPointForFilePath (event) {
-      const items = getCachedItems()
-      if (!items) return
-
-      const filePath = event.path
-      items.filter(item => item.filePath === filePath).map(item => item.refreshPoint())
-    }
+    this.provider.subscribeEditor(newEditor.onDidSave(refreshPointForFilePath))
   }
 
-  destroy () {
-    this.stop()
-    super.destroy()
+  start (options) {
+    this.provider.start(options)
   }
 
   stop () {
@@ -109,17 +117,6 @@ module.exports = class ProjectSymbols extends ProviderBase {
     }
   }
 
-  initialize () {
-    // When user manually refresh, clear cache
-    this.subscriptions.add(this.ui.onWillRefreshManually(clearCachedItems))
-  }
-
-  readTags () {
-    return new Promise(resolve => {
-      this.loadTagsTask = TagReader.getAllTags(resolve)
-    })
-  }
-
   async getItems () {
     this.stop()
     // Refresh watching target tagFile on each execution to catch-up change in outer-world.
@@ -127,21 +124,22 @@ module.exports = class ProjectSymbols extends ProviderBase {
 
     const cache = getCachedItems()
     if (cache) {
-      this.finishUpdateItems(cache)
-      return
+      this.provider.finishUpdateItems(cache)
+    } else {
+      const tags = await new Promise(resolve => {
+        this.loadTagsTask = TagReader.getAllTags(resolve)
+      })
+
+      // Better interests suggestion? I want this less noisy.
+      const kindOfInterests = 'cfm'
+      let items = tags
+        .filter(tag => kindOfInterests.includes(tag.kind))
+        .map(itemForTag)
+        .filter(item => item.point != null)
+        .sort(compareByPoint)
+      items = _.uniq(items, item => item.filePath + item.text)
+      setCachedItems(items)
+      this.provider.finishUpdateItems(items)
     }
-
-    const tags = await this.readTags()
-
-    // Better interests suggestion? I want this less noisy.
-    const kindOfInterests = 'cfm'
-    let items = tags
-      .filter(tag => kindOfInterests.includes(tag.kind))
-      .map(itemForTag)
-      .filter(item => item.point != null)
-      .sort(compareByPoint)
-    items = _.uniq(items, item => item.filePath + item.text)
-    setCachedItems(items)
-    this.finishUpdateItems(items)
   }
 }
