+++ lib/provider/scan.js
@@ -1,17 +1,6 @@
 const {Point, Range} = require('atom')
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const {scanItemsForBuffer} = require('../utils')
-
-const providerConfig = {
-  boundToSingleFile: true,
-  supportDirectEdit: true,
-  showColumnOnLineHeader: true,
-  itemHaveRange: true,
-  showSearchOption: true,
-  supportCacheItems: true,
-  useFirstQueryAsSearchTerm: true,
-  refreshOnDidStopChanging: true
-}
 
 function itemize (text, row) {
   const point = new Point(row, 0)
@@ -22,16 +11,32 @@ function itemize (text, row) {
   }
 }
 
-module.exports = class Scan extends ProviderBase {
-  constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
+module.exports = class Scan {
+  constructor () {
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        boundToSingleFile: true,
+        supportDirectEdit: true,
+        showColumnOnLineHeader: true,
+        itemHaveRange: true,
+        showSearchOption: true,
+        supportCacheItems: true,
+        useFirstQueryAsSearchTerm: true,
+        refreshOnDidStopChanging: true
+      },
+      getItems: () => this.getItems()
+    })
+  }
+
+  start (options) {
+    this.provider.start(options)
   }
 
   getItems () {
-    const {searchRegex} = this.searchOptions
-    const {buffer} = this.editor
+    const {searchRegex} = this.provider.searchOptions
+    const buffer = this.provider.editor.buffer
     const items = searchRegex ? scanItemsForBuffer(buffer, searchRegex) : buffer.getLines().map(itemize)
-    this.finishUpdateItems(items)
+    this.provider.finishUpdateItems(items)
   }
 }
