+++ lib/provider/search.js
@@ -1,13 +1,13 @@
 'use babel'
 
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const {getProjectPaths, scanItemsForFilePath, scanItemsForBuffer} = require('../utils')
 
 const Searcher = require('../searcher')
 const Path = require('path')
 const _ = require('underscore-plus')
 
-const providerConfig = {
+const Config = {
   supportDirectEdit: true,
   showColumnOnLineHeader: true,
   showProjectHeader: true,
@@ -20,50 +20,52 @@ const providerConfig = {
   refreshOnDidStopChanging: true
 }
 
-module.exports = class Search extends ProviderBase {
+module.exports = class Search {
   static searcherIsAvailable = false
 
   constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
-    this.updateItemsIfNotModified = this.updateItemsIfNotModified.bind(this)
-    this.updateItems = this.updateItems.bind(this)
-
     // FIXME: This is set to true in spec to test easily
     this.searchInOrdered = atom.inSpecMode()
+
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: Config,
+      getItems: this.getItems.bind(this),
+      onUiCreated: ui => {
+        // FIXME: Just to wait for initialization of provider.searchOptions
+        this.searcher = new Searcher(this.provider.searchOptions)
+      },
+      onDestroyed: () => {
+        if (this.searcher) {
+          this.searcher.cancel()
+          this.searcher = null
+        }
+      }
+    })
   }
 
-  getState () {
-    return this.mergeState(super.getState(), {projects: this.projects})
-  }
-
-  checkReady () {
-    if (this.reopened) {
-      console.log(this.projects)
-      return true
+  start (options = {}) {
+    this.projects = this.projects || getProjectPaths(options.currentProject ? this.provider.editor : null)
+    if (this.projects) {
+      this.provider.start(options)
     }
-
-    if (!this.projects) {
-      this.projects = getProjectPaths(this.options.currentProject ? this.editor : null)
-    }
-    return this.projects
   }
 
-  initialize () {
-    this.searcher = new Searcher(this.searchOptions)
-  }
+  // getState () {
+  //   return this.mergeState(super.getState(), {projects: this.projects})
+  // }
 
   search () {
     const projectsToSearch = this.projects.slice()
     const modifiedBuffers = atom.project.getBuffers().filter(buffer => buffer.isModified() && buffer.getPath())
-    const modifiedBuffersScanned = []
-    const {searchRegex} = this.searchOptions
+    const modifiedBuffersScanned = new Set()
+    const {searchRegex} = this.provider.searchOptions
 
     const scanBuffer = buffer => {
-      if (modifiedBuffersScanned.includes(buffer)) return
-
-      this.updateItems(scanItemsForBuffer(buffer, searchRegex))
-      modifiedBuffersScanned.push(buffer)
+      if (!modifiedBuffersScanned.includes(buffer)) {
+        this.updateItems(scanItemsForBuffer(buffer, searchRegex))
+        modifiedBuffersScanned.push(buffer)
+      }
     }
 
     let finished = 0
@@ -72,7 +74,7 @@ module.exports = class Search extends ProviderBase {
       // Append directory separator to avoid unwanted partial match.
       // make `atom` to `atom/` to avoid matches to project name like `atom-keymaps`.
       const isSameProject = buffer => buffer.getPath().startsWith(project + Path.sep)
-      modifiedBuffers.filter(isSameProject).map(scanBuffer)
+      modifiedBuffers.filter(isSameProject).forEach(scanBuffer)
 
       if (projectsToSearch.length) {
         searchNextProject()
@@ -80,12 +82,17 @@ module.exports = class Search extends ProviderBase {
       }
 
       if (finished === this.projects.length) {
-        this.finishUpdateItems()
+        this.provider.finishUpdateItems()
       }
     }
 
+    const updateItemsIfNotModified = items => {
+      items = items.filter(item => !atom.project.isPathModified(item.filePath))
+      this.updateItems(items)
+    }
+
     const searchNextProject = () => {
-      this.searcher.searchProject(projectsToSearch.shift(), this.updateItemsIfNotModified, onFinish)
+      this.searcher.searchProject(projectsToSearch.shift(), updateItemsIfNotModified, onFinish)
     }
 
     if (this.searchInOrdered) {
@@ -95,54 +102,51 @@ module.exports = class Search extends ProviderBase {
     }
   }
 
-  updateItemsIfNotModified (items) {
-    items = items.filter(item => !atom.project.isPathModified(item.filePath))
-    this.updateItems(items)
-  }
-
   updateItems (items) {
     if (this.searchInOrdered) items = _.sortBy(items, item => item.filePath)
-    super.updateItems(items)
-  }
-
-  destroy () {
-    this.searcher.cancel()
-    this.searcher = null
-    super.destroy()
+    this.provider.updateItems(items)
   }
 
   async getItems ({filePath}) {
     this.searcher.cancel()
-    const searcher = this.getConfig('searcher')
+    const searcher = this.provider.getConfig('searcher')
 
+    const executable = await this.isExecutable(searcher)
+    if (!executable) {
+      return this.provider.finishUpdateItems([])
+    }
+
+    const {searchRegex} = this.provider.searchOptions
+    if (!searchRegex) {
+      return this.provider.finishUpdateItems([])
+    }
+
+    this.searcher.setCommand(searcher)
+    if (filePath) {
+      if (atom.project.contains(filePath)) {
+        this.provider.finishUpdateItems(await scanItemsForFilePath(filePath, searchRegex))
+      } else {
+        this.provider.finishUpdateItems([])
+      }
+    } else {
+      this.search()
+    }
+  }
+
+  async isExecutable (searcherCommand) {
     if (!this.constructor.searcherIsAvailable) {
-      const status = await this.searcher.runCommandPromisified(searcher)
+      const status = await this.searcher.runCommandPromisified(searcherCommand)
       this.constructor.searcherIsAvailable = status === 0
       if (!this.constructor.searcherIsAvailable) {
         const message = [
           `atom-arrow`,
-          `- \`${searcher}\` is not available in your PATH.`,
-          `- You can choose another searcher from setting of narrow.`,
-          `- Or install/add \`${searcher}\` to PATH`
+          `- \`${searcherCommand}\` is not available in your PATH.`,
+          `- You can choose another searcherCommand from setting of narrow.`,
+          `- Or install/add \`${searcherCommand}\` to PATH`
         ].join('\n')
         atom.notifications.addWarning(message, {dismissable: true})
-        return this.finishUpdateItems([])
       }
     }
-
-    this.searcher.setCommand(searcher)
-
-    const {searchRegex} = this.searchOptions
-
-    if (!searchRegex) {
-      return this.finishUpdateItems([])
-    }
-
-    if (filePath) {
-      const items = atom.project.contains(filePath) ? await scanItemsForFilePath(filePath, searchRegex) : []
-      this.finishUpdateItems(items)
-    } else {
-      this.search()
-    }
+    return this.constructor.searcherIsAvailable
   }
 }
