+++ lib/provider/select-files.js
@@ -1,44 +1,49 @@
 // This is kind of META provider, invoked from UI.
 // Provide narrow-ui to select narrowing target files.
 const {Point} = require('atom')
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const settings = require('../settings')
 const {relativizeFilePath} = require('../utils')
 
 const queryByProviderName = {}
 
-const providerConfig = {
-  boundToSingleFile: true,
-  supportCacheItems: true,
-  supportReopen: false,
-  needRestoreEditorState: false
-}
+module.exports = class SelectFiles {
+  constructor (clientUi) {
+    this.clientUi = clientUi
 
-module.exports = class SelectFiles extends ProviderBase {
-  constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        boundToSingleFile: true,
+        supportCacheItems: true,
+        supportReopen: false,
+        needRestoreEditorState: false
+      },
+      onDestroyed: () => {
+        if (this.clientUi.isAlive()) {
+          this.clientUi.focus({autoPreview: false})
+        }
+      },
+      onItemConfirmed: this.onItemConfirmed.bind(this),
+      getItems: () => this.getItems()
+    })
+  }
+
+  start (options) {
+    this.provider.start(options)
   }
 
   static getLastQuery (providerName) {
+    let query
     if (settings.get('SelectFiles.rememberQuery')) {
       const query = queryByProviderName[providerName]
-      if (query) return query
     }
-    return ''
-  }
-
-  initialize () {
-    const {clientUi} = this.options
-    this.clientUi = clientUi
-    this.ui.onDidDestroy(() => {
-      if (clientUi.isAlive()) clientUi.focus({autoPreview: false})
-    })
+    return query || ''
   }
 
   getItems () {
     const filePaths = this.clientUi.getFilePathsForAllItems()
-    this.finishUpdateItems(
+    this.provider.finishUpdateItems(
       filePaths.map(filePath => ({
         text: relativizeFilePath(filePath),
         filePath,
@@ -47,14 +52,15 @@ module.exports = class SelectFiles extends ProviderBase {
     )
   }
 
-  async confirmed (item) {
+  // Must return undefined, when non null value was returned
+  // Ui treat it as openable editor instance and try to open it.
+  async onItemConfirmed (item) {
     if (this.clientUi.isAlive()) {
-      const queryForSelectFiles = this.ui.lastQuery
+      const queryForSelectFiles = this.provider.ui.lastQuery
       await this.clientUi.resetQueryForSelectFiles(queryForSelectFiles)
       this.clientUi.moveToItemForFilePath(item.filePath)
       queryByProviderName[this.clientUi.provider.name] = queryForSelectFiles
     }
-
-    this.ui.editor.destroy()
+    this.provider.ui.destroy()
   }
 }
