+++ lib/provider/symbols.js
@@ -1,7 +1,7 @@
 'use babel'
 
 const _ = require('underscore-plus')
-const ProviderBase = require('./provider-base')
+const Provider = require('./provider')
 const {requireFrom, getFirstCharacterPositionForBufferRow} = require('../utils')
 const TagGenerator = requireFrom('symbols-view', 'tag-generator')
 
@@ -9,59 +9,60 @@ const TagGenerator = requireFrom('symbols-view', 'tag-generator')
 // Which read tag info from file on disk.
 // So we cant update symbol unless it's saved on disk.
 // This is very exceptional provider not supportCacheItems in spite of boundToSingleFile.
+const MARKDOWN_SCOPE_NAME = ['source.gfm', 'text.md']
 
-const providerConfig = {
-  boundToSingleFile: true,
-  queryWordBoundaryOnByCurrentWordInvocation: true,
-  refreshOnDidSave: true
+function itemForTag (editor, {position, name}) {
+  const {scopeName} = editor.getGrammar()
+
+  let indentLevel
+  if (MARKDOWN_SCOPE_NAME.includes(scopeName)) {
+    // For gfm source, determine level by counting leading '#' chars( header level ).
+    const match = editor.lineTextForBufferRow(position.row).match(/#+/)
+    indentLevel = match ? match[0].length - 1 : 0
+  } else {
+    indentLevel = editor.indentationForBufferRow(position.row)
+  }
+
+  return {
+    point: getFirstCharacterPositionForBufferRow(editor, position.row),
+    text: '  '.repeat(indentLevel) + name
+  }
 }
 
-module.exports = class Symbols extends ProviderBase {
+module.exports = class Symbols {
   constructor (...args) {
-    super(...args)
-    Object.assign(this, providerConfig)
+    this.provider = Provider.create({
+      name: this.constructor.name,
+      config: {
+        boundToSingleFile: true,
+        queryWordBoundaryOnByCurrentWordInvocation: true,
+        refreshOnDidSave: true
+      },
+      onBindEditor: this.onBindEditor.bind(this),
+      getItems: this.getItems.bind(this)
+    })
+  }
+
+  start (options) {
+    this.provider.start(options)
   }
 
   onBindEditor ({newEditor}) {
-    this.items = null
-    this.subscribeEditor(newEditor.onDidSave(() => (this.items = null)))
-  }
-
-  itemForTag ({position, name}) {
-    return {
-      point: getFirstCharacterPositionForBufferRow(this.editor, position.row),
-      text: '  '.repeat(this.getLevelForRow(position.row)) + name
-    }
-  }
-
-  isMarkdownEditor (editor) {
-    const markdownScopeName = ['source.gfm', 'text.md']
-    return markdownScopeName.includes(editor.getGrammar().scopeName)
-  }
-
-  getLevelForRow (row) {
-    // For gfm source, determine level by counting leading '#' chars( use header level ).
-    if (this.isMarkdownEditor(this.editor)) {
-      const match = this.editor.lineTextForBufferRow(row).match(/#+/)
-      return match ? match[0].length - 1 : 0
-    } else {
-      return this.editor.indentationForBufferRow(row)
-    }
+    this.provider.subscribeEditor(
+      newEditor.onDidSave(() => {
+        this.items = null // invalidate cache
+      })
+    )
   }
 
   async getItems () {
-    if (this.items) {
-      this.finishUpdateItems(this.items)
-      return
+    if (!this.items) {
+      // We show full line text of symbol's line, so just care for which line have symbol.
+      const editor = this.provider.editor
+      const tags = await new TagGenerator(editor.getPath(), editor.getGrammar().scopeName).generate()
+      this.items = _.uniq(tags, tag => tag.position.row).map(tag => itemForTag(editor, tag))
     }
 
-    // We show full line text of symbol's line, so just care for which line have symbol.
-    const filePath = this.editor.getPath()
-    const {scopeName} = this.editor.getGrammar()
-
-    const tags = await new TagGenerator(filePath, scopeName).generate()
-    const itemForTag = this.itemForTag.bind(this)
-    this.items = _.uniq(tags, tag => tag.position.row).map(itemForTag)
-    this.finishUpdateItems(this.items)
+    this.provider.finishUpdateItems(this.items)
   }
 }
