+++ lib/ui.js
@@ -304,7 +304,7 @@ module.exports = class Ui {
   }
   constructor (provider, {query = ''} = {}, restoredState) {
     if (!SelectFiles) {
-      SelectFiles = require('./composition-provider/select-files')
+      SelectFiles = require('./provider/select-files')
     }
     this.provider = provider
     this.query = query
