+++ package.json
@@ -29,23 +29,7 @@
       "narrow:symbols",
       "narrow:symbols-by-current-word",
       "narrow:toggle-search-start-by-double-click",
-      "narrow:activate-package",
-      "narrow:x-atom-scan",
-      "narrow:x-atom-scan-by-current-word",
-      "narrow:x-fold",
-      "narrow:x-fold-by-current-word",
-      "narrow:x-git-diff-all",
-      "narrow:x-project-symbols",
-      "narrow:x-project-symbols-by-current-word",
-      "narrow:x-scan",
-      "narrow:x-scan-by-current-word",
-      "narrow:x-search",
-      "narrow:x-search-by-current-word",
-      "narrow:x-search-by-current-word-without-focus",
-      "narrow:x-search-current-project",
-      "narrow:x-search-current-project-by-current-word",
-      "narrow:x-symbols",
-      "narrow:x-symbols-by-current-word"
+      "narrow:activate-package"
     ],
     "atom-text-editor.vim-mode-plus-search": [
       "vim-mode-plus-user:narrow:scan",
